server/requirements.txt
fastapi
uvicorn
python-dotenv
spotipy
pandas
scikit-learn
numpy
requests
python-multipart
